// frontend/api/models/User.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { UserRole } from './UserRole';
export type User = {
    id?: number;
    role?: UserRole;
    email?: string;
    nickname?: string;
    phone?: string;
    avatar?: string;
    is_active?: boolean;
    created_at?: string;
};